Update solana trigger e2e tests for guaranteed delivery
