+++ system-tests/tests/smoke/cre/solana_capability_test.go
@@ -369,11 +369,16 @@ func ExecuteSolanaLogTriggerTest(t *testing.T, tenv *configuration.TestEnvironme
 		tenv, testLogger, workflowName, &workflowConfig,
 		workflowFileLocation)
 
-	emitCtx, emitCancel := context.WithCancel(t.Context())
-	defer emitCancel()
+	singleAckFound, stopACKLogScans := startTriggerEventACKLogWatch(t, testLogger)
+	defer stopACKLogScans()
+
 	ticker := time.NewTicker(10 * time.Second)
+	emitCtx, emitCancel := context.WithCancel(t.Context())
 	go func() {
-		defer ticker.Stop()
+		defer func() {
+			emitCancel()
+			ticker.Stop()
+		}()
 		for {
 			select {
 			case <-emitCtx.Done():
@@ -381,7 +386,7 @@ func ExecuteSolanaLogTriggerTest(t *testing.T, tenv *configuration.TestEnvironme
 			case <-ticker.C:
 				slot, err := callCreateLog(emitCtx, solChain, logReadTestProgramID, expectedU64Value)
 				if err == nil {
-					t.Logf("Log read test event triggered at slot: %d", slot)
+					testLogger.Info().Uint64("slot", slot).Msg("Log read test event triggered")
 				}
 			}
 		}
@@ -390,7 +395,11 @@ func ExecuteSolanaLogTriggerTest(t *testing.T, tenv *configuration.TestEnvironme
 	expectedLogTriggerMessage := fmt.Sprintf("TestEvent received: str_val=%s u64_value=%d", expectedStrVal, expectedU64Value)
 	t_helpers.WatchWorkflowLogs(t, testLogger, userLogsCh, baseMessageCh,
 		t_helpers.WorkflowEngineInitErrorLog, expectedLogTriggerMessage,
-		5*time.Minute, t_helpers.WithUserLogWorkflowID(workflowID))
+		4*time.Minute, t_helpers.WithUserLogWorkflowID(workflowID))
+	emitCancel()
+	testLogger.Info().Msgf("Found expected user log: '%s'", expectedLogTriggerMessage)
+
+	requireTriggerEventACKLog(t, testLogger, singleAckFound)
 }
 
 func callCreateLog(ctx context.Context, solChain *solana.Blockchain, programID solgo.PublicKey, value uint64) (slot uint64, err error) {
@@ -529,11 +538,16 @@ func ExecuteSolanaLogTriggerCPITest(t *testing.T, tenv *configuration.TestEnviro
 		tenv, testLogger, workflowName, &workflowConfig,
 		workflowFileLocation)
 
-	emitCtx, emitCancel := context.WithCancel(t.Context())
-	defer emitCancel()
+	singleAckFound, stopACKLogScans := startTriggerEventACKLogWatch(t, testLogger)
+	defer stopACKLogScans()
+
 	ticker := time.NewTicker(10 * time.Second)
+	emitCtx, emitCancel := context.WithCancel(t.Context())
 	go func() {
-		defer ticker.Stop()
+		defer func() {
+			emitCancel()
+			ticker.Stop()
+		}()
 		for {
 			select {
 			case <-emitCtx.Done():
@@ -541,7 +555,7 @@ func ExecuteSolanaLogTriggerCPITest(t *testing.T, tenv *configuration.TestEnviro
 			case <-ticker.C:
 				slot, err := triggerLogReadTestCPIEvent(emitCtx, solChain, logReadTestProgramID, expectedU64Value)
 				if err == nil {
-					t.Logf("Log read test CPI event triggered at slot: %d", slot)
+					testLogger.Info().Uint64("slot", slot).Msg("Log read test CPI event triggered")
 				}
 			}
 		}
@@ -550,5 +564,9 @@ func ExecuteSolanaLogTriggerCPITest(t *testing.T, tenv *configuration.TestEnviro
 	expectedLogTriggerMessage := fmt.Sprintf("TestEvent CPI received: str_val=%s u64_value=%d", expectedStrVal, expectedU64Value)
 	t_helpers.WatchWorkflowLogs(t, testLogger, userLogsCh, baseMessageCh,
 		t_helpers.WorkflowEngineInitErrorLog, expectedLogTriggerMessage,
-		5*time.Minute, t_helpers.WithUserLogWorkflowID(workflowID))
+		4*time.Minute, t_helpers.WithUserLogWorkflowID(workflowID))
+	emitCancel()
+	testLogger.Info().Msgf("Found expected user log: '%s'", expectedLogTriggerMessage)
+
+	requireTriggerEventACKLog(t, testLogger, singleAckFound)
 }
